Further improved Linux support and fixed a few bugs along the way.

+++ FastCG/include/FastCG/FlyCameraController.h
@@ -35,7 +35,7 @@ namespace FastCG
 		void OnUpdate(float time, float deltaTime) override;
 
 	private:
-		glm::vec2 mLastMousePosition{0, 0};
+		glm::uvec2 mLastMousePosition{0, 0};
 		bool mRightMouseButtonPressed{false};
 		float mWalkSpeed{1};
 		float mTurnSpeed{1};
